+++ packages/template-babel/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.3](https://github.com/void-aurora/toolkit/compare/@void-aurora/template-babel@1.2.2...@void-aurora/template-babel@1.2.3) (2020-03-14)
+
+**Note:** Version bump only for package @void-aurora/template-babel
+
 ## [1.2.2](https://github.com/void-aurora/toolkit/compare/@void-aurora/template-babel@1.2.1...@void-aurora/template-babel@1.2.2) (2020-03-07)
 
 **Note:** Version bump only for package @void-aurora/template-babel